Added user validation for deleting and updating a note

+++ src/resolvers/mutation.js
@@ -18,7 +18,17 @@ module.exports = {
             author: new mongoose.Types.ObjectId(user.id)
         });
     },
-    deleteNote: async (parent, {id}, {models}) => {
+    deleteNote: async (parent, {id}, {models, user}) => {
+        if(!user){
+            throw new AuthenticationError('You must be signed in to create a note')
+        }
+
+        const note = await models.Note.findById(id);
+        
+        if(note && String(note.author) != user.id){
+            throw new ForbiddenError("You don't have permission to delete the note");
+        }
+        
         try {
             await models.Note.findOneAndDelete({_id: id});
             return true;
@@ -27,6 +37,15 @@ module.exports = {
         }
     },
     updateNote: async (parent, {content, id}, {models}) => {
+        if(!user){
+            throw new AuthenticationError('You must be signed in to create a note')
+        }
+
+        const note = await models.Note.findById(id);
+        
+        if(note && String(note.author) != user.id){
+            throw new ForbiddenError("You don't have permission to delete the note");
+        }
         return await models.Note.findOneAndUpdate(
             {
                 _id: id
